moving the logic into the handler.
